Module types::metered_channel

Structs

Permit

A newtype for an mpsc::Permit which allows us to inject gauge accounting in the case the permit is dropped w/o sending

Receiver

An mpsc::Receiver with an [IntGauge] counting the number of currently queued items.

ReceiverStream

Stream API Wrappers! A wrapper around crate::metered_channel::Receiver that implements [Stream].

Sender

An mpsc::Sender with an [IntGauge] counting the number of currently queued items.

Functions

channel

Constructor Similar to mpsc::channel, channel creates a pair of Sender and Receiver