Function types::proof::treebits::disk_write_order

pub fn disk_write_order(node: u64, levels_to_collapse: u32) -> u64

Defines an order for the nodes optimized for writing to disk. In this order all of the nodes with the same level(node)/levels_to_collapse and a common parent will share the same Disk Write Order. The Disk Write Order of a node will be greater than that of any of it's children. In this tree:

     3
    /  \
   /    \
  1      5
 / \    / \
0   2  4   6

With levels_to_collapse=2, 0, 1, 2 have a DWO of 0; 4,5,6 a DWO of 1. 3 will have a DWO of 4 to account for the fact that the nodes with DWOs of 2 and 3 will have parents that have a DWO of 4.