Crate types

Modules

error

Macros

bail

ensure

Structs

Batch

BatchDigest

BatchDigestProto

Proto wrapper for Narwhal type

BatchMessage

BatchMessageProto

Proto wrapper for Narwhal type

BatchProto

Proto wrapper for Narwhal type

BincodeEncodedPayload

A bincode encoded payload. This is intended to be used in the short-term while we don’t have good protobuf definitions for Narwhal types

BlockError

Certificate

CertificateDigest

CertificateDigestProto

Proto wrapper for Narwhal type

ClientBatchRequest

CollectionError

Proto wrapper for Narwhal type

CollectionRetrievalResult

Empty

Empty message for when we don’t have anything to return

GetCollectionsRequest

GetCollectionsResponse

Header

HeaderBuilder

Builder for Header.

HeaderDigest

PrimaryToPrimaryClient

The primary-to-primary interface

PrimaryToPrimaryServer

The primary-to-primary interface

PrimaryToWorkerClient

The primary-to-worker interface

PrimaryToWorkerServer

The primary-to-worker interface

TransactionProto

Proto wrapper for Narwhal type

TransactionsClient

TransactionsServer

ValidatorClient

The consensus to mempool interface for validator actions.

ValidatorServer

The consensus to mempool interface for validator actions.

Vote

VoteDigest

WorkerToPrimaryClient

The worker-to-primary interface

WorkerToPrimaryServer

The worker-to-primary interface

WorkerToWorkerClient

The worker-to-worker interface

WorkerToWorkerServer

The worker-to-worker interface

Enums

BlockErrorType

CollectionErrorType

Proto wrapper for Narwhal type

HeaderBuilderError

Error type for HeaderBuilder

PrimaryMessage

PrimaryWorkerMessage

The messages sent by the primary to its workers.

RetrievalResult

WorkerMessage

The message exchanged between workers.

Traits

PrimaryToPrimary

Generated trait containing gRPC methods that should be implemented for use with PrimaryToPrimaryServer.

PrimaryToWorker

Generated trait containing gRPC methods that should be implemented for use with PrimaryToWorkerServer.

Transactions

Generated trait containing gRPC methods that should be implemented for use with TransactionsServer.

Validator

Generated trait containing gRPC methods that should be implemented for use with ValidatorServer.

WorkerToPrimary

Generated trait containing gRPC methods that should be implemented for use with WorkerToPrimaryServer.

WorkerToWorker

Generated trait containing gRPC methods that should be implemented for use with WorkerToWorkerServer.

Type Definitions

BlockResult

Round

The round number.

Transaction